Parse ffprobe's output as UTF-8
Fixes files with UTF-8 tags being read as line noise. ffprobe is
guaranteed to output UTF-8, however whether the tag's encoding really
was UTF-8, was correctly guessed as ISO-8859-1 or incorrectly guessed
as ISO-8859-1 is a possible issue.

Update the test vector to allow the test suite to verify UTF-8-ness.